+++ app/src/main/java/com/iabrmv/mindmaps/repository/MindmapRepository.kt
@@ -1,30 +1,28 @@
-package com.iabrmv.mindmaps.business
+package com.iabrmv.mindmaps.repository
 
-import com.iabrmv.mindmaps.business.model.Mindmap
-import com.iabrmv.mindmaps.database.*
+import com.iabrmv.mindmaps.data.business.Mindmap
+import com.iabrmv.mindmaps.data.database.*
 import io.realm.Realm
 import io.realm.kotlin.where
 
-class MindmapManager(val realm: Realm) {
+class MindmapRepository(val realm: Realm) {
 
     fun loadMindmap(id: Int) = realm
         .where<MindmapEntity>()
         .equalTo("id", id)
         .findFirst()?.toBusinessModel()
 
-    fun loadMindmaps(): List<Mindmap> = realm
+    fun loadMindmapsFromDB(): List<Mindmap> = realm
         .where<MindmapEntity>()
         .findAll()
-        .map {
-            it.toBusinessModel()
-        }
+        .map { it.toBusinessModel() }
+        .sortedByDescending { it.lastEditedTimeMillis }
 
-    fun saveMindmap(mindmap: Mindmap) {
+    fun saveMindmapToDB(mindmap: Mindmap) {
         realm.executeTransaction { realm ->
             realm.copyToRealmOrUpdate(mindmap.toEntity())
         }
     }
-
 
     fun clear() {
         realm.executeTransaction { realm ->
